Make wo requests from base and involvementApis
